+++ components/components.ts
@@ -31,6 +31,9 @@ export { default as Checkbox } from './checkbox'
 export type { Props as CollapseProps, Item as CollapseItem } from './collapse'
 export { default as Collapse } from './collapse'
 
+export type { Props as ColorPickerProps } from './colorpicker'
+export { default as ColorPicker } from './colorpicker'
+
 export type { Props as CountdownProps } from './countdown'
 export { default as Countdown } from './countdown'