+++ docs/guide/changelog.md
@@ -4,6 +4,14 @@
 
 对于新功能、新组件、`bug` 修复以及文档更新，您可以向 `main` 分支创建拉取请求或通过右下角邮箱地址联系我
 
+## 2.1.0 <Tag color="volcano" size="small">soon</Tag>
+
+- 新增 [颜色选择器 ColorPicker](https://themusecatcher.github.io/vue-amazing-ui/guide/components/colorpicker.html) 组件
+
+## 2.0.8 <Tag color="volcano" size="small">2024-01-09</Tag>
+
+- 组件库及文档代码优化
+
 ## 2.0.7 <Tag color="volcano" size="small">2024-01-07</Tag>
 
 - 优化并更新 [轮播图 Carousel](https://themusecatcher.github.io/vue-amazing-ui/guide/components/carousel.html) 组件，图片类型 `Image` 新增 `link` `target` 属性，支持：设置图片跳转链接；设置如何打开跳转链接
@@ -450,7 +458,6 @@
 - 新增 布局 Layout 组件
 - 新增 菜单 Menu 组件
 - 新增 穿梭框 Transfer 组件
-- 新增 颜色选择器 ColorPicker 组件
 <!-- - 时间轴 Timeline 组件，新增水平时间轴
 - 新增 时间选择器 TimePicker 组件
 - 重构级联选择 Cascader 组件